Implement Desktop and Windows

+++ src/drivers/mouse.cpp
@@ -66,7 +66,7 @@ uint32_t MouseDriver::HandleInterrupt(uint32_t esp) {
     offset = (offset + 1) % 3;
     if(offset == 0) {
         if(buffer[1] != 0 || buffer[2] != 0) {
-            handler->OnMouseMove(buffer[1], -buffer[2]);
+            handler->OnMouseMove((int8_t)buffer[1], -((int8_t)buffer[2]));
 
         }
